requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyJWT==2.8.0 
bcrypt==4.1.2 
python-dotenv==1.0.0
SQLAlchemy==2.0.36
gunicorn==21.2.0
PyMuPDF==1.23.8
python-docx==1.1.0
syllables==1.0.2 
nltk==3.8.1
spacy==3.7.2
transformers==4.36.2 
torch==2.1.0 
matplotlib==3.8.2
seaborn==0.13.0